# backend/docs/swagger.yaml
definitions:
...
  common.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: 请求参数错误
        type: string
      timestamp:
        example: "2025-08-29T10:00:00Z"
        type: string
    type: object
  common.PaginationResponse:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      has_next:
        example: true
        type: boolean
      has_previous:
        example: false
        type: boolean
      message:
        example: success
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      timestamp:
        example: "2025-08-29T10:00:00Z"
        type: string
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  common.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: success
        type: string
      timestamp:
        example: "2025-08-29T10:00:00Z"
        type: string
    type: object
...
  github_com_lighthought_app-maker_backend_internal_models.ChangePasswordRequest:
    properties:
      new_password:
        example: newpassword123
        minLength: 6
        type: string
      old_password:
        example: oldpassword123
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  github_com_lighthought_app-maker_backend_internal_models.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_lighthought_app-maker_backend_internal_models.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        example: 3600
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.UserInfo'
    type: object
...
  github_com_lighthought_app-maker_backend_internal_models.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      username:
        example: username
        maxLength: 20
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  github_com_lighthought_app-maker_backend_internal_models.UpdateProfileRequest:
    properties:
      email:
        example: newemail@example.com
        type: string
      username:
        example: newusername
        maxLength: 20
        minLength: 3
        type: string
    type: object
...
  github_com_lighthought_app-maker_backend_internal_models.UpdateUserSettingsRequest:
    properties:
      auto_go_next:
        description: 自动进入下一阶段配置
        example: false
        type: boolean
      default_ai_model:
        example: glm-4.6
        type: string
      default_api_token:
        example: sk-...
        type: string
      default_cli_tool:
        enum:
        - claude-code
        - qwen-code
        - gemini
        example: claude-code
        type: string
      default_model_api_url:
        example: https://open.bigmodel.cn/api/anthropic
        type: string
      default_model_provider:
        enum:
        - ollama
        - zhipu
        - anthropic
        - openai
        - vllm
        example: zhipu
        type: string
    type: object
  github_com_lighthought_app-maker_backend_internal_models.UserInfo:
    properties:
      created_at:
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: varchar(50)
        type: string
      role:
        example: user
        type: string
      status:
        example: active
        type: string
      username:
        example: username
        type: string
    type: object
  github_com_lighthought_app-maker_backend_internal_models.UserSettingsResponse:
    properties:
      auto_go_next:
        description: 自动进入下一阶段配置
        example: false
        type: boolean
      default_ai_model:
        example: glm-4.6
        type: string
      default_api_token:
        description: 敏感信息，前端可能需要脱敏显示
        example: sk-***
        type: string
      default_cli_tool:
        example: claude-code
        type: string
      default_model_api_url:
        example: https://open.bigmodel.cn/api/anthropic
        type: string
      default_model_provider:
        example: zhipu
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录并获取访问令牌
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 用户登录
      tags:
      - 用户管理
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌
      parameters:
      - description: 刷新令牌
        in: query
        name: refresh_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 刷新令牌
      tags:
      - 用户管理
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账户
      parameters:
      - description: 注册请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 用户注册
      tags:
      - 用户管理
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: 获取用户列表（需要管理员权限）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/common.PaginationResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取用户列表
      tags:
      - 用户管理
  /api/v1/users/{user_id}:
    delete:
      consumes:
      - application/json
      description: 删除指定用户（需要管理员权限）
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 删除用户
      tags:
      - 用户管理
  /api/v1/users/change-password:
    post:
      consumes:
      - application/json
      description: 修改当前用户的密码
      parameters:
      - description: 修改密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 修改密码
      tags:
      - 用户管理
  /api/v1/users/logout:
    post:
      consumes:
      - application/json
      description: 用户登出并清除会话
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 用户登出
      tags:
      - 用户管理
  /api/v1/users/profile:
    get:
      consumes:
      - application/json
      description: 获取当前用户的档案信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.UserInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取用户档案
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 更新当前用户的档案信息
      parameters:
      - description: 更新档案请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新用户档案
      tags:
      - 用户管理
  /api/v1/users/settings:
    get:
      consumes:
      - application/json
      description: 获取当前用户的开发设置（CLI工具、模型配置等）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.UserSettingsResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取用户设置
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 更新当前用户的开发设置（CLI工具、模型配置等）
      parameters:
      - description: 更新设置请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.UpdateUserSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新用户设置
      tags:
      - 用户管理
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"